Create a gas strategies module in nucypher.utilities
